refactor: work 詳細ページコンポーネントをディレクトリ構造に再編成
_internal/components/ 配下の各コンポーネントを
独立したディレクトリに移動し、共通定義を抽出

- BackButton/: 新規作成
- ThumbnailGallery/: コンポーネント簡素化、スタイル分離
- WorkDescriptions/: スタイルを styles.css.ts に分離
- definitions.ts: 共通の型定義を抽出
- services.ts: 画像処理ユーティリティを移動
- _index.css.ts: 削除（各コンポーネントに分散）
