+++ src/pages/works/[id]/_internal/BackButton/styles.css.ts
@@ -1,0 +1,10 @@
+import { style } from '@vanilla-extract/css';
+
+import { vars } from '../../../../../foundation/styles/theme.css';
+
+export const backButton = style({
+	display: 'flex',
+	alignItems: 'center',
+	gap: '8px',
+	font: vars.text.en.description,
+});
