+++ src/pages/works/[id]/_internal/WorkDescriptions/styles.css.ts
@@ -1,6 +1,6 @@
 import { createVar, fallbackVar, globalStyle, style } from '@vanilla-extract/css';
 
-import { BreakPoints, vars } from '../../../../../styles/theme.css';
+import { BreakPoints, vars } from '../../../../../foundation/styles/theme.css';
 
 export const root = style({
 	display: 'flex',
